Simulate a path from a model

The named verb for the package's simulate.stochastic_model() method. It exists so that the public surface reads as one family of aci_* verbs. It is the same call, with the same arguments and the same draws, and the S3 method stays registered on the stats::simulate() generic, so simulate(model) keeps working.

Usage

aci_simulate(object, ...)

Arguments

object

A stochastic_model or cgns_model object.

...

Passed to the simulate() method: nsim, seed, t_end, dt, ic, burn_in, keep_hidden, keep_noise.

Value

Whatever the method returns: an aci_sim object, or a list of them when nsim exceeds one.

See also

Examples

m <- aci_dyad_model()
aci_simulate(m, seed = 1, t_end = 2, dt = 0.01)
#> $obs
#> <obs_traj> k = 1, N+1 = 201, dt = 0.01, span [0, 2]
#> 
#> $hidden
#>                [,1]
#>   [1,]  2.000000000
#>   [2,]  2.020940184
#>   [3,]  2.169374213
#>   [4,]  2.304774115
#>   [5,]  2.247583863
#>   [6,]  1.989459145
#>   [7,]  2.207530107
#>   [8,]  2.241056449
#>   [9,]  2.257710239
#>  [10,]  2.213478588
#>  [11,]  2.216360295
#>  [12,]  2.148663163
#>  [13,]  2.130350610
#>  [14,]  2.023652144
#>  [15,]  1.817383622
#>  [16,]  1.850878064
#>  [17,]  1.928508105
#>  [18,]  1.817795674
#>  [19,]  1.607359867
#>  [20,]  1.577530661
#>  [21,]  1.469682676
#>  [22,]  1.184551066
#>  [23,]  1.063681539
#>  [24,]  0.870915777
#>  [25,]  0.702715146
#>  [26,]  0.460236798
#>  [27,]  0.508740962
#>  [28,]  0.342166740
#>  [29,]  0.047822620
#>  [30,] -0.058076792
#>  [31,] -0.153803292
#>  [32,] -0.374987247
#>  [33,] -0.790808342
#>  [34,] -0.976243590
#>  [35,] -1.036573097
#>  [36,] -1.153448076
#>  [37,] -1.261012575
#>  [38,] -1.294402527
#>  [39,] -1.494748496
#>  [40,] -1.459972940
#>  [41,] -1.534298997
#>  [42,] -1.534459471
#>  [43,] -1.495469646
#>  [44,] -1.531384316
#>  [45,] -1.675167730
#>  [46,] -1.611366862
#>  [47,] -1.856731527
#>  [48,] -1.948634693
#>  [49,] -2.007997746
#>  [50,] -2.056258847
#>  [51,] -1.981019122
#>  [52,] -1.993256184
#>  [53,] -1.975796680
#>  [54,] -2.000890457
#>  [55,] -2.041522966
#>  [56,] -1.981893865
#>  [57,] -1.878504880
#>  [58,] -2.133023760
#>  [59,] -2.085226825
#>  [60,] -2.052502613
#>  [61,] -2.099099626
#>  [62,] -2.005572444
#>  [63,] -2.049834574
#>  [64,] -2.081278611
#>  [65,] -1.997957005
#>  [66,] -1.827012909
#>  [67,] -1.798795587
#>  [68,] -1.839013331
#>  [69,] -1.951489782
#>  [70,] -1.979070890
#>  [71,] -2.066579484
#>  [72,] -2.088206468
#>  [73,] -2.043940254
#>  [74,] -2.122084093
#>  [75,] -1.849679531
#>  [76,] -1.825484561
#>  [77,] -1.701568174
#>  [78,] -1.920024359
#>  [79,] -1.833346747
#>  [80,] -1.952345946
#>  [81,] -1.846804123
#>  [82,] -1.792914057
#>  [83,] -1.818990832
#>  [84,] -1.671384035
#>  [85,] -1.727912766
#>  [86,] -1.770503501
#>  [87,] -1.855235992
#>  [88,] -1.906335929
#>  [89,] -1.796558192
#>  [90,] -1.737938538
#>  [91,] -1.622567474
#>  [92,] -1.647317639
#>  [93,] -1.594648068
#>  [94,] -1.556316584
#>  [95,] -1.686098122
#>  [96,] -1.495210271
#>  [97,] -1.471091163
#>  [98,] -1.384273877
#>  [99,] -1.277172256
#> [100,] -1.270371813
#> [101,] -1.287883529
#> [102,] -1.184875858
#> [103,] -1.275947039
#> [104,] -1.064665066
#> [105,] -1.089309995
#> [106,] -0.910116343
#> [107,] -0.745627482
#> [108,] -0.725925191
#> [109,] -0.658383201
#> [110,] -0.751040747
#> [111,] -0.708762837
#> [112,] -0.596162903
#> [113,] -0.577850953
#> [114,] -0.614464904
#> [115,] -0.672406195
#> [116,] -0.667336428
#> [117,] -0.551510195
#> [118,] -0.591149630
#> [119,] -0.593976006
#> [120,] -0.713828932
#> [121,] -0.654623188
#> [122,] -0.514143947
#> [123,] -0.354995395
#> [124,] -0.266009322
#> [125,] -0.445711068
#> [126,] -0.389178683
#> [127,] -0.335410432
#> [128,] -0.363499922
#> [129,] -0.338981559
#> [130,] -0.417989922
#> [131,] -0.341659751
#> [132,] -0.366059272
#> [133,] -0.514586724
#> [134,] -0.541734790
#> [135,] -0.395967907
#> [136,] -0.419647096
#> [137,] -0.336723800
#> [138,] -0.231877508
#> [139,] -0.221632623
#> [140,] -0.246873501
#> [141,] -0.289461300
#> [142,] -0.204908541
#> [143,] -0.300487970
#> [144,] -0.264986956
#> [145,] -0.282797175
#> [146,] -0.497990907
#> [147,] -0.626398950
#> [148,] -0.521666139
#> [149,] -0.528393849
#> [150,] -0.435649600
#> [151,] -0.234765022
#> [152,] -0.076223959
#> [153,]  0.001883607
#> [154,]  0.049870455
#> [155,]  0.040338712
#> [156,]  0.207870628
#> [157,]  0.276157672
#> [158,]  0.166837647
#> [159,]  0.160171901
#> [160,] -0.022920267
#> [161,] -0.033175713
#> [162,] -0.282466729
#> [163,] -0.139779017
#> [164,] -0.192779251
#> [165,] -0.224827293
#> [166,] -0.230645613
#> [167,] -0.158270764
#> [168,] -0.079897174
#> [169,] -0.012916874
#> [170,] -0.060133379
#> [171,] -0.186185248
#> [172,] -0.214175309
#> [173,] -0.175886580
#> [174,] -0.247949394
#> [175,] -0.244429305
#> [176,] -0.350801667
#> [177,] -0.340598976
#> [178,] -0.316738341
#> [179,] -0.320827090
#> [180,] -0.327912816
#> [181,] -0.142983883
#> [182,] -0.060103532
#> [183,]  0.058262533
#> [184,] -0.028292812
#> [185,] -0.005918460
#> [186,]  0.116591848
#> [187,]  0.116887585
#> [188,] -0.090034954
#> [189,] -0.049893687
#> [190,] -0.234242709
#> [191,] -0.307950505
#> [192,] -0.167024286
#> [193,] -0.097513658
#> [194,]  0.018909646
#> [195,]  0.056773551
#> [196,]  0.052072635
#> [197,] -0.033144959
#> [198,]  0.134432342
#> [199,]  0.145344294
#> [200,]  0.080857223
#> [201,]  0.174381496
#> 
#> $noise
#> $B
#>           [,1]     [,2]     [,3]       [,4]      [,5]     [,6]      [,7]
#> [1,] 0.4094018 1.688873 1.586588 -0.3309078 -2.285236 2.497662 0.6670662
#>           [,8]        [,9]     [,10]      [,11]     [,12]      [,13]     [,14]
#> [1,] 0.5413273 -0.01339952 0.5101084 -0.1643758 0.4206946 -0.4002467 -1.370208
#>          [,15]    [,16]      [,17]    [,18]     [,19]       [,20]     [,21]
#> [1,] 0.9878383 1.519745 -0.3087406 -1.25329 0.6422413 -0.04470914 -1.733218
#>           [,22]      [,23]      [,24]     [,25]    [,26]     [,27]     [,28]
#> [1,] 0.00213186 -0.6303003 -0.3409686 -1.156572 1.803142 -0.331132 -1.605513
#>          [,29]     [,30]      [,31]     [,32]      [,33]     [,34]       [,35]
#> [1,] 0.1971934 0.2631756 -0.9858267 -2.888921 -0.6404817 0.5705076 -0.05972328
#>            [,36]     [,37]     [,38]    [,39]        [,40]     [,41]    [,42]
#> [1,] -0.09817874 0.5608207 -1.186459 1.096777 -0.005344028 0.7073107 1.034108
#>          [,43]      [,44]    [,45]     [,46]      [,47]      [,48]     [,49]
#> [1,] 0.2234804 -0.8787076 1.162965 -2.000165 -0.5447907 -0.2556707 -0.166121
#>         [,50]     [,51]     [,52]       [,53]      [,54]     [,55]    [,56]
#> [1,] 1.020464 0.1362219 0.4071676 -0.06965481 -0.2476643 0.6955508 1.146228
#>          [,57]     [,58]     [,59]      [,60]     [,61]      [,62]      [,63]
#> [1,] -2.403096 0.5727396 0.3747244 -0.4252677 0.9510128 -0.3892372 -0.2843307
#>          [,64]    [,65]     [,66]     [,67]     [,68]     [,69]      [,70]
#> [1,] 0.8574098 1.719627 0.2700549 -0.422184 -1.189113 -0.331033 -0.9398293
#>           [,71]     [,72]      [,73]    [,74]     [,75]    [,76]     [,77]
#> [1,] -0.2589326 0.3943792 -0.8518571 2.649167 0.1560117 1.130207 -2.289124
#>          [,78]     [,79]     [,80]     [,81]      [,82]    [,83]      [,84]
#> [1,] 0.7410012 -1.316245 0.9198037 0.3981302 -0.4075286 1.324259 -0.7012317
#>           [,85]     [,86]      [,87]    [,88]     [,89]    [,90]      [,91]
#> [1,] -0.5806143 -1.001072 -0.6681786 0.945185 0.4337021 1.005159 -0.3901187
#>          [,92]     [,93]     [,94]    [,95]     [,96]    [,97]     [,98]
#> [1,] 0.3763703 0.2441649 -1.426257 1.778429 0.1344477 0.765599 0.9551367
#>           [,99]     [,100]    [,101]    [,102]   [,103]     [,104]   [,105]
#> [1,] -0.0505657 -0.3058154 0.8936737 -1.047298 1.971337 -0.3836321 1.654145
#>        [,106]     [,107]    [,108]    [,109]    [,110]   [,111]     [,112]
#> [1,] 1.512213 0.08296573 0.5672209 -1.024548 0.3230065 1.043612 0.09907849
#>          [,113]     [,114]      [,115]   [,116]     [,117]     [,118]   [,119]
#> [1,] -0.4541369 -0.6557819 -0.03592242 1.069161 -0.4839749 -0.1210101 -1.29414
#>         [,120]   [,121]   [,122]    [,123]    [,124]    [,125]    [,126]
#> [1,] 0.4943128 1.307902 1.497041 0.8147027 -1.869789 0.4820295 0.4561356
#>          [,127]    [,128]    [,129]    [,130]    [,131]    [,132]     [,133]
#> [1,] -0.3534003 0.1704895 -0.864036 0.6792308 -0.327101 -1.569082 -0.3674508
#>        [,134]     [,135]  [,136]    [,137]      [,138]     [,139]     [,140]
#> [1,] 1.364435 -0.3342814 0.73275 0.9465856 0.004398704 -0.3523223 -0.5296955
#>         [,141]    [,142]    [,143]     [,144]    [,145]   [,146]    [,147]
#> [1,] 0.7395892 -1.063457 0.2462108 -0.2894994 -2.264889 -1.40885 0.9160193
#>         [,148]    [,149]   [,150]   [,151]    [,152]    [,153]     [,154]
#> [1,] -0.191279 0.8032832 1.887474 1.473881 0.6772685 0.3799627 -0.1927984
#>        [,155]    [,156]    [,157]     [,158]   [,159]     [,160]    [,161]
#> [1,] 1.577892 0.5962341 -1.173577 -0.1556425 -1.91891 -0.1952588 -2.592328
#>        [,162]    [,163]     [,164]     [,165]    [,166]    [,167]   [,168]
#> [1,] 1.314002 -0.635543 -0.4299788 -0.1693183 0.6122182 0.6783402 0.567952
#>          [,169]    [,170]     [,171]    [,172]     [,173]      [,174]    [,175]
#> [1,] -0.5725426 -1.363291 -0.3887222 0.2779141 -0.8230811 -0.06884093 -1.167662
#>            [,176]    [,177]     [,178]    [,179]   [,180]    [,181]   [,182]
#> [1,] -0.008309014 0.1288554 -0.1458756 -0.163911 1.763552 0.7625865 1.111431
#>         [,183]    [,184]   [,185]      [,186]    [,187]    [,188]    [,189]
#> [1,] -0.923207 0.1643418 1.154825 -0.05652142 -2.129361 0.3448458 -1.904955
#>          [,190]   [,191]    [,192]   [,193]    [,194]     [,195]     [,196]
#> [1,] -0.8111702 1.324004 0.6156368 1.091669 0.3066049 -0.1101588 -0.9243128
#>        [,197]    [,198]     [,199]    [,200]
#> [1,] 1.592914 0.0450106 -0.7151284 0.8652231
#> 
#> $W
#>            [,1]      [,2]       [,3]     [,4]      [,5]       [,6]      [,7]
#> [1,] -0.6264538 0.1836433 -0.8356286 1.595281 0.3295078 -0.8204684 0.4874291
#>           [,8]      [,9]      [,10]    [,11]     [,12]      [,13]   [,14]
#> [1,] 0.7383247 0.5757814 -0.3053884 1.511781 0.3898432 -0.6212406 -2.2147
#>         [,15]       [,16]       [,17]     [,18]     [,19]     [,20]     [,21]
#> [1,] 1.124931 -0.04493361 -0.01619026 0.9438362 0.8212212 0.5939013 0.9189774
#>          [,22]      [,23]     [,24]     [,25]       [,26]      [,27]     [,28]
#> [1,] 0.7821363 0.07456498 -1.989352 0.6198257 -0.05612874 -0.1557955 -1.470752
#>           [,29]     [,30]   [,31]      [,32]     [,33]       [,34]    [,35]
#> [1,] -0.4781501 0.4179416 1.35868 -0.1027877 0.3876716 -0.05380504 -1.37706
#>           [,36]    [,37]      [,38]    [,39]     [,40]      [,41]      [,42]
#> [1,] -0.4149946 -0.39429 -0.0593134 1.100025 0.7631757 -0.1645236 -0.2533617
#>          [,43]     [,44]      [,45]      [,46]    [,47]     [,48]      [,49]
#> [1,] 0.6969634 0.5566632 -0.6887557 -0.7074952 0.364582 0.7685329 -0.1123462
#>          [,50]     [,51]      [,52]     [,53]     [,54]    [,55]  [,56]
#> [1,] 0.8811077 0.3981059 -0.6120264 0.3411197 -1.129363 1.433024 1.9804
#>           [,57]     [,58]     [,59]      [,60]    [,61]    [,62]     [,63]
#> [1,] -0.3672215 -1.044135 0.5697196 -0.1350546 2.401618 -0.03924 0.6897394
#>           [,64]      [,65]     [,66]     [,67]    [,68]     [,69]    [,70]
#> [1,] 0.02800216 -0.7432732 0.1887923 -1.804959 1.465555 0.1532533 2.172612
#>          [,71]      [,72]     [,73]      [,74]     [,75]     [,76]      [,77]
#> [1,] 0.4755095 -0.7099464 0.6107264 -0.9340976 -1.253633 0.2914462 -0.4432919
#>            [,78]      [,79]      [,80]      [,81]      [,82]    [,83]     [,84]
#> [1,] 0.001105352 0.07434132 -0.5895209 -0.5686687 -0.1351786 1.178087 -1.523567
#>          [,85]     [,86]  [,87]      [,88]     [,89]     [,90]    [,91]
#> [1,] 0.5939462 0.3329504 1.0631 -0.3041839 0.3700188 0.2670988 -0.54252
#>         [,92]    [,93]     [,94]    [,95]     [,96]     [,97]      [,98]
#> [1,] 1.207868 1.160403 0.7002136 1.586833 0.5584864 -1.276592 -0.5732654
#>          [,99]     [,100]     [,101]     [,102]     [,103]    [,104]     [,105]
#> [1,] -1.224613 -0.4734006 -0.6203667 0.04211587 -0.9109216 0.1580288 -0.6545846
#>        [,106]    [,107]    [,108]    [,109]   [,110]     [,111]     [,112]
#> [1,] 1.767287 0.7167075 0.9101742 0.3841854 1.682176 -0.6357365 -0.4616447
#>        [,113]     [,114]     [,115]     [,116]     [,117]     [,118]    [,119]
#> [1,] 1.432282 -0.6506964 -0.2073807 -0.3928079 -0.3199929 -0.2791133 0.4941883
#>          [,120]     [,121]   [,122]     [,123]     [,124]     [,125]    [,126]
#> [1,] -0.1773305 -0.5059575 1.343039 -0.2145794 -0.1795565 -0.1001907 0.7126663
#>          [,127]      [,128]     [,129]     [,130]     [,131]     [,132]
#> [1,] -0.0735644 -0.03763417 -0.6816605 -0.3242703 0.06016044 -0.5888945
#>         [,133]    [,134]    [,135]   [,136]     [,137]     [,138]     [,139]
#> [1,] 0.5314962 -1.518394 0.3065579 -1.53645 -0.3009761 -0.5282799 -0.6520948
#>           [,140]    [,141]   [,142]    [,143]     [,144]   [,145]    [,146]
#> [1,] -0.05689678 -1.914359 1.176583 -1.664972 -0.4635304 -1.11592 -0.750819
#>        [,147]     [,148]    [,149]    [,150]    [,151]      [,152]     [,153]
#> [1,] 2.087167 0.01739562 -1.286301 -1.640606 0.4501871 -0.01855983 -0.3180684
#>          [,154]   [,155]    [,156]   [,157]     [,158]    [,159]   [,160]
#> [1,] -0.9293621 -1.48746 -1.075192 1.000029 -0.6212667 -1.384427 1.869291
#>         [,161]     [,162]   [,163]    [,164]    [,165]   [,166]    [,167]
#> [1,] 0.4251004 -0.2386471 1.058483 0.8864227 -0.619243 2.206102 -0.255027
#>         [,168]     [,169]    [,170]   [,171]    [,172]    [,173]      [,174]
#> [1,] -1.424495 -0.1443996 0.2075383 2.307978 0.1058024 0.4569988 -0.07715294
#>          [,175]      [,176]    [,177]   [,178]   [,179]   [,180]    [,181]
#> [1,] -0.3340008 -0.03472603 0.7876396 2.075245 1.027392 1.207908 -1.231323
#>         [,182]    [,183]   [,184]    [,185]     [,186]   [,187]    [,188]
#> [1,] 0.9838956 0.2199248 -1.46725 0.5210227 -0.1587546 1.464587 -0.766082
#>          [,189]     [,190]    [,191]    [,192]     [,193]    [,194]    [,195]
#> [1,] -0.4302118 -0.9261095 -0.177104 0.4020118 -0.7317482 0.8303732 -1.208083
#>         [,196]   [,197]    [,198]    [,199]     [,200]
#> [1,] -1.047984 1.441158 -1.015847 0.4119747 -0.3810761
#> 
#> attr(,"class")
#> [1] "noise_store"
#> 
#> $model
#> <cgns_model> 'dyad[p1] y->x': k = 1 observed, l = 1 hidden
#> 
#> $meta
#> $meta$perfect_model
#> [1] TRUE
#> 
#> $meta$scheme
#> [1] "euler_maruyama"
#> 
#> $meta$dt
#> [1] 0.01
#> 
#> $meta$burn_in
#> [1] 0
#> 
#> $meta$burn_interval
#> [1] 0 0
#> 
#> 
#> attr(,"class")
#> [1] "aci_sim"